Extract flexible layout functions to separate layout class.
